# ai-portfolio-intelligence/apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AIConfigureRequest": {
        "properties": {
          "api_key": {
            "minLength": 10,
            "title": "Api Key",
            "type": "string"
          },
          "model": {
            "default": "gemini-2.5-flash",
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "AIConfigureRequest",
        "type": "object"
      },
      "AIScheduleSettings": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "midday_time": {
            "default": "12:30",
            "title": "Midday Time",
            "type": "string"
          },
          "morning_time": {
            "default": "09:30",
            "title": "Morning Time",
            "type": "string"
          },
          "night_time": {
            "default": "20:00",
            "title": "Night Time",
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "AIScheduleSettings",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScheduledAnalyzeRequest": {
        "properties": {
          "period": {
            "title": "Period",
            "type": "string"
          }
        },
        "required": [
          "period"
        ],
        "title": "ScheduledAnalyzeRequest",
        "type": "object"
      },
...
      "ThesisUpdateRequest": {
        "properties": {
          "break_triggers": {
            "items": {
              "type": "string"
            },
            "title": "Break Triggers",
            "type": "array"
          },
          "key_assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Key Assumptions",
            "type": "array"
          },
          "thesis": {
            "minLength": 10,
            "title": "Thesis",
            "type": "string"
          }
        },
        "required": [
          "thesis"
        ],
        "title": "ThesisUpdateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Read-only AI portfolio intelligence and research API.",
    "title": "Portfolio Analyzer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai/analyze-portfolio": {
      "post": {
        "operationId": "analyze_portfolio_ai_analyze_portfolio_post",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze Portfolio",
        "tags": [
          "ai-research"
        ]
      }
    },
    "/ai/analyze-stock/{symbol}": {
      "post": {
        "operationId": "analyze_stock_ai_analyze_stock__symbol__post",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "con_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Con Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze Stock",
        "tags": [
          "ai-research"
        ]
      }
    },
...
    "/ai/configure": {
      "post": {
        "operationId": "configure_ai_ai_configure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConfigureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Configure Ai Ai Configure Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Configure Ai",
        "tags": [
          "ai-research"
        ]
      }
    },
    "/ai/schedule": {
      "get": {
        "operationId": "get_schedule_ai_schedule_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Schedule Ai Schedule Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "ai-research"
        ]
      },
      "put": {
        "operationId": "update_schedule_ai_schedule_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIScheduleSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Schedule Ai Schedule Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Schedule",
        "tags": [
          "ai-research"
        ]
      }
    },
    "/ai/scheduled-analyze": {
      "post": {
        "description": "Trigger a mock or real scheduled daily slot analysis (Morning, Midday, Night).",
        "operationId": "trigger_scheduled_analysis_ai_scheduled_analyze_post",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Scheduled Analysis",
        "tags": [
          "ai-research"
        ]
      }
    },
    "/ai/status": {
      "get": {
        "operationId": "ai_status_ai_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Ai Status Ai Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ai Status",
        "tags": [
          "ai-research"
        ]
      }
    },
    "/ai/thesis/{symbol}": {
      "get": {
        "operationId": "read_thesis_ai_thesis__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Thesis Ai Thesis  Symbol  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Thesis",
        "tags": [
          "ai-research"
        ]
      },
      "put": {
        "operationId": "write_thesis_ai_thesis__symbol__put",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThesisUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Write Thesis Ai Thesis  Symbol  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Write Thesis",
        "tags": [
          "ai-research"
        ]
      }
    },
...
  }
}